Move pool stuff into a pool package
