+++ src/main/java/io/vertx/core/http/impl/pool/ConnectionListener.java
@@ -13,7 +13,7 @@
  *
  * You may elect to redistribute this code under either of these licenses.
  */
-package io.vertx.core.http.impl;
+package io.vertx.core.http.impl.pool;
 
 import io.netty.channel.Channel;
